+++ tests/Feature/RememberALessonTest.php
@@ -8,7 +8,7 @@ test('Remember a lesson that God is learning me', function () {
         message: 'The lesson was not added.'
     );
 
-    $renouncedLessons = shell_exec('php index.php lessons:renounce');
+    $renouncedLessons = shell_exec('php index.php lessons:repeat');
     $this->assertStringContainsString(
         needle: 'Dit is een les die God mij heeft geleerd.',
         haystack: $renouncedLessons,